+++ examples/sklearn-logreg-mnist/sklearn_example/client_app.py
@@ -1,56 +1,25 @@
+"""sklearn_example: A Flower / scikit-learn app."""
+
 import warnings
 
-import numpy as np
-from flwr_datasets import FederatedDataset
-from flwr_datasets.partitioner import IidPartitioner
 from sklearn.linear_model import LogisticRegression
 from sklearn.metrics import log_loss
+from sklearn_example.task import (
+    get_model_parameters,
+    load_data,
+    set_initial_params,
+    set_model_params,
+)
 
 from flwr.client import Client, ClientApp, NumPyClient
-from flwr.common import Context, NDArrays
-
-
-def get_model_parameters(model: LogisticRegression) -> NDArrays:
-    """Returns the parameters of a sklearn LogisticRegression model."""
-    if model.fit_intercept:
-        params = [
-            model.coef_,
-            model.intercept_,
-        ]
-    else:
-        params = [
-            model.coef_,
-        ]
-    return params
-
-
-def set_model_params(model: LogisticRegression, params: NDArrays) -> LogisticRegression:
-    """Sets the parameters of a sklean LogisticRegression model."""
-    model.coef_ = params[0]
-    if model.fit_intercept:
-        model.intercept_ = params[1]
-    return model
-
-
-def set_initial_params(model: LogisticRegression):
-    """Sets initial parameters as zeros Required since model params are uninitialized
-    until model.fit is called.
-
-    But server asks for initial parameters from clients at launch. Refer to
-    sklearn.linear_model.LogisticRegression documentation for more information.
-    """
-    n_classes = 10  # MNIST has 10 classes
-    n_features = 784  # Number of features in dataset
-    model.classes_ = np.array([i for i in range(10)])
-
-    model.coef_ = np.zeros((n_classes, n_features))
-    if model.fit_intercept:
-        model.intercept_ = np.zeros((n_classes,))
+from flwr.common import Context
 
 
 # Define Flower client
 class MnistClient(NumPyClient):
-    def __init__(self, model, X_train, X_test, y_train, y_test):
+    def __init__(
+        self, model, X_train, X_test, y_train, y_test
+    ):  # pylint: disable=R0913
         self.model = model
         self.X_train = X_train
         self.X_test = X_test
@@ -76,40 +45,14 @@ class MnistClient(NumPyClient):
         return loss, len(self.X_test), {"accuracy": accuracy}
 
 
-fds = None  # Cache FederatedDataset
-
-
-def load_data(partition_id: int, num_partitions: int):
-    # Only initialize `FederatedDataset` once
-    global fds
-    if fds is None:
-        partitioner = IidPartitioner(num_partitions=num_partitions)
-        fds = FederatedDataset(
-            dataset="ylecun/mnist",
-            partitioners={"train": partitioner},
-        )
-
-    dataset = fds.load_partition(partition_id, "train").with_format("numpy")
-    X, y = dataset["image"].reshape((len(dataset), -1)), dataset["label"]
-    # Split the on edge data: 80% train, 20% test
-    X_train, X_test = X[: int(0.8 * len(X))], X[int(0.8 * len(X)) :]
-    y_train, y_test = y[: int(0.8 * len(y))], y[int(0.8 * len(y)) :]
-
-    return X_train, X_test, y_train, y_test
-
-
 def client_fn(context: Context) -> Client:
-    """Construct a Client that will be run in a ClientApp.
-
-    You can use settings in `context.run_config` to parameterize the
-    construction of your Client. You could use the `context.node_config` to, for
-    example, indicate which dataset to load (e.g accesing the partition-id).
-    """
+    """Construct a Client that will be run in a ClientApp."""
 
     # Read the node_config to fetch data partition associated to this node
     partition_id = context.node_config["partition-id"]
     num_partitions = context.node_config["num-partitions"]
 
+    # Load train and test data
     X_train, X_test, y_train, y_test = load_data(partition_id, num_partitions)
 
     # Create LogisticRegression Model
@@ -119,10 +62,12 @@ def client_fn(context: Context) -> Client:
         warm_start=True,  # prevent refreshing weights when fitting
     )
 
-    # Setting initial parameters, akin to model.compile for keras models
+    # Set initial parameters, akin to model.compile for keras models
     set_initial_params(model)
 
+    # Return Client instance
     return MnistClient(model, X_train, X_test, y_train, y_test).to_client()
 
 
+# Create ClientApp
 app = ClientApp(client_fn=client_fn)
