+++ examples/sklearn-logreg-mnist/sklearn_example/server_app.py
@@ -1,37 +1,16 @@
+"""sklearn_example: A Flower / scikit-learn app."""
+
 from typing import Dict
 
-import numpy as np
+# import numpy as np
 from flwr_datasets import FederatedDataset
 from sklearn.linear_model import LogisticRegression
 from sklearn.metrics import log_loss
+from sklearn_example.task import set_initial_params, set_model_params
 
 from flwr.common import Context, NDArrays
 from flwr.server import ServerApp, ServerAppComponents, ServerConfig
 from flwr.server.strategy import FedAvg
-
-
-def set_model_params(model: LogisticRegression, params: NDArrays) -> LogisticRegression:
-    """Sets the parameters of a sklean LogisticRegression model."""
-    model.coef_ = params[0]
-    if model.fit_intercept:
-        model.intercept_ = params[1]
-    return model
-
-
-def set_initial_params(model: LogisticRegression):
-    """Sets initial parameters as zeros Required since model params are uninitialized
-    until model.fit is called.
-
-    But server asks for initial parameters from clients at launch. Refer to
-    sklearn.linear_model.LogisticRegression documentation for more information.
-    """
-    n_classes = 10  # MNIST has 10 classes
-    n_features = 784  # Number of features in dataset
-    model.classes_ = np.array([i for i in range(10)])
-
-    model.coef_ = np.zeros((n_classes, n_features))
-    if model.fit_intercept:
-        model.intercept_ = np.zeros((n_classes,))
 
 
 def fit_round(server_round: int) -> Dict:
@@ -62,23 +41,21 @@ def get_evaluate_fn():
 
 
 def server_fn(context: Context) -> ServerAppComponents:
-    """Construct components that set the ServerApp behaviour.
+    """Construct components that set the ServerApp behaviour."""
 
-    You can use settings in `context.run_config` to parameterize the
-    construction of all elements (e.g the strategy or the number of rounds)
-    wrapped in the returned ServerAppComponents object.
-    """
+    # Read from config
+    num_rounds = context.run_config["num-server-rounds"]
 
+    # Define the strategy
     strategy = FedAvg(
         min_available_clients=2,
         evaluate_fn=get_evaluate_fn(),
         on_fit_config_fn=fit_round,
     )
-    # Construct ServerConfig
-    num_rounds = context.run_config["num-server-rounds"]
     config = ServerConfig(num_rounds=num_rounds)
 
     return ServerAppComponents(strategy=strategy, config=config)
 
 
+# Create ServerApp
 app = ServerApp(server_fn=server_fn)
